requirements.txt
fastapi
uvicorn
pymongo[srv]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
jinja2
email-validator
Pillow
pytesseract
slowapi
httpx

#optional for oauth
authlib
httpx